Initial commit, showing how the directory stucture will work.
Octant tools will live in the top directory as a .py file.  E.g., grid.py contains all of the grid utilities.  These are imported into the octant namespace using import/from statements.  There are (presently) two notable exceptions.

The extern directory was created to store code that is used by octant (so that we reduce the number of dependencies), but is not part of the official codebase.  This keeps the copywrite statements/test/etc. in place.

Second is the io module.  This is included separately since it needs to run in order to figure out what to import.  This may change in the future, but I like having it farmed off to octant.io for now.

The other subdirectories should be moved to the top of the namespace by concatenating the files they contain. 


git-svn-id: http://octant.googlecode.com/svn/trunk@2 443bacc9-c644-0410-9ac0-e35418122bc4